requirements.txt
mlagents==0.10.0
PyYAML==3.13
torch==1.4.0